backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.4.2
pydantic-core==2.10.1
python-multipart==0.0.6
aiofiles==23.2.1
cryptography==41.0.7
python-dotenv==1.0.0
scikit-learn==1.3.2
joblib==1.3.2
pandas==2.1.4